Fixing imports and moving files
