+++ src/dashboard/sections/YourDesk/ControlButton/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import './styles.css';
-import { Spinner } from '../../../../../components/layoutPieces';
+import { Spinner } from '../../../../components';
 
 const propTypes = {
   isActive: PropTypes.bool,
