+++ src/dashboard/sections/YourDesk/GenericTab/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import './styles.css';
-import Spinner from '../../../../../components/layoutPieces/Spinner';
+import Spinner from '../../../../components/layoutPieces/Spinner';
 import MetricsFormatter from './MetricsFormatter';
 import Ranking from '../Ranking';
 
@@ -44,8 +44,7 @@ function GenericTab({ tab, error, metrics, ranks, map, tabTitle }) {
   const hasMetrics = Object.keys(metrics).length;
   const hasRank = ranks.length;
   const hasRight = Object.keys(map).length || ranks.length > 1;
-  console.log('metrics', metrics);
-  console.log('tab', tab);
+
   return (
     <div className="GenericTab-main">
       <div className="GenericTab-upper">
