--- src/dashboard/sections/YourDesk/YourDesk/index.js
@@ -1,4 +1,0 @@
-export { default as ControlButton } from './ControlButton';
-// export { default as OpenCasesDetail } from './openCasesDetail';
-export { default as GenericTab } from './GenericTab';
-export { default as Ranking } from './Ranking';
